saythanks-mobile-testing/requirements.txt
selenium==3.141.0
pytest==6.2.5
pytest-html==2.1.1
pytest-xdist==2.5.0
pytest-cov==3.0.0
requests==2.25.1
pillow==8.4.0
beautifulsoup4==4.10.0
lxml==4.6.5
py==1.11.0
appium-python-client==1.3.0